Update the explorer for the merged parser
The merge brings four changes the explorer has to account for.

- The parser's fields are reordered to a cache line and narrowed, `chunkBase`
  is gone (the span builder takes the base as an argument and is `static`),
  the held UTF-8 tail is a `UInt64` field rather than the buffer's last eight
  bytes, and the escape scratch is a reserved region past `bufferCapacity`
  written whole in one unaligned store. `chunk-in`, `buffering` and `escapes`
  said the old thing; they now say this one, and the pinned assembly is
  re-extracted against the merged binary (`consumeStringRun` 720 -> 853,
  which is `recordInlineChunk` back inline where it belongs).
- Blocks are `StreamBlock` -- a `ManagedBuffer` owning its capacity -- and
  each array holds its own count of the filling block, so a shared block is
  written past rather than copied and the commit path has no uniqueness check
  on it at all. `collections` said "the first commit after a snapshot copies
  one block", which is no longer true of any block size.
- Element initial values are one leaked template per schema copied through
  `_streamCopyInitialize`, replacing the two initial-value closure forms.
- Skipped escapes reject a raw control byte and hand a non-ASCII selector to
  UTF-8 validation, and the windowed object arm checks the key-to-colon gap
  and defers a value the chunk cut.

The doc's last chapter described the in-place open element as shipped; the
last commit on the branch reverted it. It is retitled to carry both verdicts
and closes with what the revert left standing, so the graveyard stops
claiming a design the source contradicts. Two orphaned comment blocks the
revert left in `StreamSchema` are removed.

`StorageTraces` now takes a snapshot partway through the fill and holds it,
recording per step whether the filling block is still the object the snapshot
captured -- read off the shipped values, and failing the trace build if an
append ever copies it.

+++ Sources/StreamParsingSiteTool/Trace/SinkTraces.swift
@@ -349,7 +349,7 @@ enum SinkTraces {
     parser.state = .skipping
     parser.depth = startDepth
     parser.containers = seed
-    parser.skipEndDepth = startDepth - 1
+    parser.skipEndDepth = UInt8(startDepth - 1)
     var sink = RecordingSink()
     let shippedEnd = try bytes.withUnsafeBytes { raw -> Int in
       try parser.consumeSkipRun(
